Fix the benchmark clock and the chat reply, and fold Generate into Chats
The NPU and CPU both reading exactly 2.00 ms was not a coincidence and not
the NPU doing nothing: `run()` divided nanoseconds by a million into a Long,
so every sample was a whole millisecond. A graph taking 1.6 ms on one and
2.4 ms on the other reported 2 ms on both, and the speedup came out as
exactly 1.00x. It returns fractional milliseconds now, the table keeps three
decimals, and a ratio near 1.00 is no longer presented as a finding -- the
report says which of the three causes it is: the clock could not resolve the
gap, the two overlap inside their own spread, or the APU genuinely does not
help on this graph.

Two real bugs in the text engine, found by reading the official litert-samples
apps rather than guessing the API. Replies were rendered with the message's
`toString()` instead of extracting its `Content.Text` parts, which puts a
data class's debug representation into the conversation. And the callback
API it used is not the one four Google samples use: `sendMessageAsync(text)`
returns a `Flow<Message>`.

Chats now carry what a reply cost. Time to first token is measured
separately from decode, because on an NPU that is where prefill and any
just-in-time compilation land and it is usually most of a short reply. Token
counts are labelled as estimates throughout, since LiteRT-LM exposes none
and the tokenizer is inside the bundle; the estimate splits ASCII from the
rest, because the length/4 rule Google's samples use is a Latin-text figure
that reads twice as optimistic on Cyrillic -- and the context gauge is the
thing it would mislead.

Reasoning goes in a grey panel that expands and collapses under an arrow.
Inline it buries the answer; dropped it takes away the one thing that
explains a wrong one.

Generate is no longer a tab. Prompt, settings and result were split across
two screens, so the picture appeared somewhere other than where it was asked
for and the seed that produced it was on a third. It is the image chat now.
Conversations can be deleted and sort by last message.

Inference settings expose what LiteRT-LM actually has: context window,
temperature, top-K, top-P. Not KV cache precision -- that is baked into the
bundle at conversion time, which is what q4 and q8 in the file names mean,
and the engine takes no dtype. Settings says so rather than offering a
switch wired to nothing.

+++ app/src/main/java/dev/neuroforge/imaging/TiledUpscaler.kt
@@ -75,7 +75,7 @@ class TiledUpscaler(
         inBuf, pixelBuf, normalization,
       )
       session.writeInput(0, inBuf)
-      totalMillis += session.run().toDouble()
+      totalMillis += session.run()
       val out = session.readOutput(0)
 
       val outW = tile.srcW * grid.scale
